requirements_dev_optional.txt
# optional library requirements
bsddb3==6.2.6; sys_platform != 'win32'
lmdb==0.97; sys_platform != 'win32'
# optional library requirements for Jupyter
ipytree==0.1.3
# optional library requirements for services
# don't let pyup change pinning for azure-storage-blob, need to pin to older
# version to get compatibility with azure storage emulator on appveyor
azure-storage-blob==2.0.1 # pyup: ignore
redis==3.3.8
pymongo==3.9.0
# optional test requirements
tox==3.14.0
coverage==5.0.3
coveralls==1.11.1
flake8==3.8.3
pytest-cov==2.7.1
pytest-doctestplus==0.4.0
pytest-remotedata==0.3.2
h5py==2.10.0
moto>=1.3.14; python_version > '3.6'
flask
s3fs==0.5.0; python_version > '3.6'
# all async features in v3
pytest-trio ; python_version >= '3.6'
trio ; python_version >= '3.6'
redio ; python_version >= '3.7' and sys_platform != 'win32'
xarray ; python_version >= '3.8'
netCDF4 ; python_version >= '3.8'
